black_box::utils::system_utils

Struct SystemUtils

pub struct SystemUtils;

Struct for SystemUtils, Contains the common utilities abstracting time, file etc.

Author

K Vishwanath

Implementations

impl SystemUtils

Implementation of SystemUtils Struct

Author

K Vishwanath

const LARGE_FILE_SIZE: u64 = 104_857_600u64

100 MB

pub fn get_timestamp() -> String

Returns a timestamp in the format yyyy-MM-dd_HH:mm:ss,SSS

pub fn is_file_large(path: &PathBuf) -> Result<bool, AppError>

It is assumed that the passed path is a file. If using RustRover or any IDE where the code snippet fs::metadata(path) highlights path as red you can ignore this.

Arguments

path - Address of PathBuf

Example

let file_path = PathBuf::from("./sample/path");
let is_large = SystemUtils::is_file_large(file_path);

pub fn form_path(paths: &[&str]) -> Result<PathBuf, AppError>

This forms a path from the project root directory

Arguments

paths - Address of array of string slice

Example

let formed_path = SystemUtils::form_path(&["sampleFolder1","sampleFolder2"]);